Refatorando Back e Front

+++ Back/usuario/deletar.php
@@ -1,12 +1,14 @@
 <?php
     include '../main.php';
 
+    if (!isset($_SESSION['user'])) {
+        http_response_code(401);
+        exit();
+    }
 
     $id = $_POST['id'];
 
     $stmt = $conn->prepare('DELETE FROM usuarios WHERE id = ?');
-    $stmt->execute([
-        $id,
-    ])
+    $stmt->execute([ $id ])
 
 ?>
